requirements.txt
Flask
python-dotenv
google-generativeai
docxtpl
Pillow